==> Utilities/8. Compile_table2.R <==
################################################
#### code to compile Table 2  for EUVME ##
################################################

# get table 1
  Regiontab1 <- subset(table1, table1$Ecoregion %in% Region_ID)

  
# get fishable domain in the region
  Fishdom <- s_EUFootp

# get fished area in the region
  Mobile_fish <- st_intersection(s_New_mobile,s_reg) ; Mobile_fish <- st_make_valid(Mobile_fish)
  Static_fish <- st_intersection(s_New_static,s_reg) ; Static_fish <- st_make_valid(Static_fish)

# get fishing SAR in the region
  
  vmsreg          <- readRDS(paste(pathdir,paste("1-Input data/VMS data repository/SARdata_2009_2022.rds",sep=""),sep="/"))
  vmsreg$c_square <- vmsreg$c.square
  #vmsreg <- read.csv(paste(pathdir_nogit, paste("VMS data repository/adhoc_VMS_datacall",datacallyear_VMS,".csv",sep=""),sep="/"))
  nam_fished      <- c(paste("SAR_total",newyear_fished,sep="_"))
  indexcol_fished <- which(names(vmsreg) %in% nam_fished) 
  vmsreg$mobeff   <- rowMeans(vmsreg[indexcol_fished],na.rm=T)
  vmsreg$mobeff[is.na(vmsreg$mobeff)] <- 0
  Regiontab1 <- cbind(Regiontab1, vmsreg[match(Regiontab1$csquares,vmsreg$c_square), c("mobeff")])
  colnames(Regiontab1)[ncol(Regiontab1)] <- "mobeff"
  Regiontab1 <- subset(Regiontab1,Regiontab1$cat =="de4_8")
  
# make the table
  tab2 <- as.data.frame(matrix(data=NA,nrow = 14, ncol= 6))

# calculate percentage overlap with fishable domain
  FD_scen11 <- st_intersection(s_scen11_dpt,Fishdom); FD_scen11 <- st_make_valid(FD_scen11)
  tab2[4,2] <- round(sum(st_area(FD_scen11))/sum(st_area(Fishdom)) * 100,digits = 1)
  
  FD_scen12 <- st_intersection(s_scen12_dpt,Fishdom); FD_scen12 <- st_make_valid(FD_scen12)
  tab2[4,3] <- round(sum(st_area(FD_scen12))/sum(st_area(Fishdom)) * 100,digits = 1)
  
  FD_scen21 <- st_intersection(s_scen21_dpt,Fishdom); FD_scen21 <- st_make_valid(FD_scen21)
  tab2[4,4] <- round(sum(st_area(FD_scen21))/sum(st_area(Fishdom)) * 100,digits = 1)
  
  FD_scen22 <- st_intersection(s_scen22_dpt,Fishdom); FD_scen22 <- st_make_valid(FD_scen22)
  tab2[4,5] <- round(sum(st_area(FD_scen22))/sum(st_area(Fishdom)) * 100,digits = 1)
  
  FD_scen23 <- st_intersection(s_scen23_dpt,Fishdom); FD_scen23 <- st_make_valid(FD_scen23)
  tab2[4,6] <- round(sum(st_area(FD_scen23))/sum(st_area(Fishdom)) * 100,digits = 1)

# calculate percentage overlap with fishing closures
  closure_overlap_scen11 <- st_intersection(s_scen11_dpt, s_clos_eu_d) %>% st_make_valid()
  tab2[5,2] <- round(sum(st_area(closure_overlap_scen11))/sum(st_area(s_scen11_dpt)) * 100,digits = 1)
    
  closure_overlap_scen12 <- st_intersection(s_scen12_dpt, s_clos_eu_d) %>% st_make_valid()
  tab2[5,3] <- round(sum(st_area(closure_overlap_scen12))/sum(st_area(s_scen12_dpt)) * 100,digits = 1)
  
  closure_overlap_scen21 <- st_intersection(s_scen21_dpt, s_clos_eu_d) %>% st_make_valid()
  tab2[5,4] <- round(sum(st_area(closure_overlap_scen21))/sum(st_area(s_scen21_dpt)) * 100,digits = 1)
  
  closure_overlap_scen22 <- st_intersection(s_scen22_dpt, s_clos_eu_d) %>% st_make_valid()
  tab2[5,5] <- round(sum(st_area(closure_overlap_scen22))/sum(st_area(s_scen22_dpt)) * 100,digits = 1)
  
  closure_overlap_scen23 <- st_intersection(s_scen23_dpt, s_clos_eu_d) %>% st_make_valid()
  tab2[5,6] <- round(sum(st_area(closure_overlap_scen23))/sum(st_area(s_scen23_dpt)) * 100,digits = 1)
  
    
# Number of VME polygons and their average areal extent (size)
  tab2[6,2] <- paste(nrow(s_scen11_dpt)," (",round(mean(st_area(s_scen11_dpt)/10^6),digits=1)," km<sup>2</sup>)",sep="")
  tab2[6,3] <- paste(nrow(s_scen12_dpt)," (",round(mean(st_area(s_scen12_dpt)/10^6),digits=1)," km<sup>2</sup>)",sep="")
  tab2[6,4] <- paste(nrow(s_scen21_dpt)," (",round(mean(st_area(s_scen21_dpt)/10^6),digits=1)," km<sup>2</sup>)",sep="")
  tab2[6,5] <- paste(nrow(s_scen22_dpt)," (",round(mean(st_area(s_scen22_dpt)/10^6),digits=1)," km<sup>2</sup>)",sep="")
  tab2[6,6] <- paste(nrow(s_scen23_dpt)," (",round(mean(st_area(s_scen23_dpt)/10^6),digits=1)," km<sup>2</sup>)",sep="")

# Number of large VME polygons in upper 25th percentile of the size distribution nd their average areal extent (size)
  sc11areas <- sort(st_area(s_scen11_dpt)/10^6)
  sc12areas <- sort(st_area(s_scen12_dpt)/10^6)
  sc21areas <- sort(st_area(s_scen21_dpt)/10^6)
  sc22areas <- sort(st_area(s_scen22_dpt)/10^6)
  sc23areas <- sort(st_area(s_scen23_dpt)/10^6)
  
  sc11p25 <- which(as.numeric(cumsum(sort(st_area(s_scen11_dpt)/10^6))/sum(st_area(s_scen11_dpt)/10^6))>0.75)
  sc12p25 <- which(as.numeric(cumsum(sort(st_area(s_scen12_dpt)/10^6))/sum(st_area(s_scen12_dpt)/10^6))>0.75)
  sc21p25 <- which(as.numeric(cumsum(sort(st_area(s_scen21_dpt)/10^6))/sum(st_area(s_scen21_dpt)/10^6))>0.75)
  sc22p25 <- which(as.numeric(cumsum(sort(st_area(s_scen22_dpt)/10^6))/sum(st_area(s_scen22_dpt)/10^6))>0.75)
  sc23p25 <- which(as.numeric(cumsum(sort(st_area(s_scen23_dpt)/10^6))/sum(st_area(s_scen23_dpt)/10^6))>0.75)

  tab2[7,2] <- paste(length(sc11p25)," (",round(mean(sc11areas[sc11p25]),digits=1)," km<sup>2</sup>)",sep="")
  tab2[7,3] <- paste(length(sc12p25)," (",round(mean(sc12areas[sc12p25]),digits=1)," km<sup>2</sup>)",sep="")
  tab2[7,4] <- paste(length(sc21p25)," (",round(mean(sc21areas[sc21p25]),digits=1)," km<sup>2</sup>)",sep="")
  tab2[7,5] <- paste(length(sc22p25)," (",round(mean(sc22areas[sc22p25]),digits=1)," km<sup>2</sup>)",sep="")
  tab2[7,6] <- paste(length(sc23p25)," (",round(mean(sc23areas[sc23p25]),digits=1)," km<sup>2</sup>)",sep="")

  
# Total area closed
  
  tab2[8,2] <- paste(round(sum(st_area(s_scen11_dpt)/10^6),digits=1)," km<sup>2</sup>",sep="")
  tab2[8,3] <- paste(round(sum(st_area(s_scen12_dpt)/10^6),digits=1)," km<sup>2</sup>",sep="")
  tab2[8,4] <- paste(round(sum(st_area(s_scen21_dpt)/10^6),digits=1)," km<sup>2</sup>",sep="")
  tab2[8,5] <- paste(round(sum(st_area(s_scen22_dpt)/10^6),digits=1)," km<sup>2</sup>",sep="")
  tab2[8,6] <- paste(round(sum(st_area(s_scen23_dpt)/10^6),digits=1)," km<sup>2</sup>",sep="")
  

  # overlap between static fishing effort and vme polygons
# not available

# overlap between static fishing area and vme polygons
  SF_scen11 <- st_intersection(s_scen11_dpt,Static_fish); SF_scen11 <- st_make_valid(SF_scen11)
  tab2[12,2] <- round(sum(st_area(SF_scen11))/sum(st_area(Static_fish)) * 100,digits = 1)
  
  SF_scen12 <- st_intersection(s_scen12_dpt,Static_fish); SF_scen12 <- st_make_valid(SF_scen12)
  tab2[12,3] <- round(sum(st_area(SF_scen12))/sum(st_area(Static_fish)) * 100,digits = 1)
  
  SF_scen21 <- st_intersection(s_scen21_dpt,Static_fish); SF_scen21 <- st_make_valid(SF_scen21)
  tab2[12,4] <- round(sum(st_area(SF_scen21))/sum(st_area(Static_fish)) * 100,digits = 1)
  
  SF_scen22 <- st_intersection(s_scen22_dpt,Static_fish); SF_scen22 <- st_make_valid(SF_scen22)
  tab2[12,5] <- round(sum(st_area(SF_scen22))/sum(st_area(Static_fish)) * 100,digits = 1)
  
  SF_scen23 <- st_intersection(s_scen23_dpt,Static_fish); SF_scen23 <- st_make_valid(SF_scen23)
  tab2[12,6] <- round(sum(st_area(SF_scen23))/sum(st_area(Static_fish)) * 100,digits = 1)

# overlap between mobile fishing effort and vme polygons
  load(paste(pathdir,paste("2-Data processing/VME_polygons",datacallyear,sep="_"),"sce11_quarter_csq_grid.RData",sep="/"))
  sce11 <- subset(sce11,sce11@data$csquares %in% Regiontab1$csquares)
  if(nrow(sce11)!=0){
  sce11@data$area_sqkm <- area(sce11)/10^6
  sce11 <- cbind(sce11, Regiontab1[match(sce11$csquares,Regiontab1$csquares), c("mobeff")])
  colnames(sce11@data)[ncol(sce11@data)] <- "mobeff"
  tab2[13,2] <- round(sum(sce11@data$area_sqkm * sce11@data$mobeff,na.rm=T) / sum(Regiontab1$area_sqkm * Regiontab1$mobeff,na.rm=T) * 100,digits =1)
  } else {
    tab2[13,2] <- "NA"
  }
  
  load(paste(pathdir,paste("2-Data processing/VME_polygons",datacallyear,sep="_"),"sce12_quarter_csq_grid.RData",sep="/"))
  sce12 <- subset(sce12,sce12@data$csquares %in% Regiontab1$csquares)
  if(nrow(sce12)!=0){
  sce12@data$area_sqkm <- area(sce12)/10^6
  sce12 <- cbind(sce12, Regiontab1[match(sce12$csquares,Regiontab1$csquares), c("mobeff")])
  colnames(sce12@data)[ncol(sce12@data)] <- "mobeff"
  tab2[13,3] <- round(sum(sce12@data$area_sqkm * sce12@data$mobeff,na.rm=T) / sum(Regiontab1$area_sqkm * Regiontab1$mobeff,na.rm=T) * 100,digits =1)
  } else {
    tab2[13,3] <- "NA"
  }
  
  load(paste(pathdir,paste("2-Data processing/VME_polygons",datacallyear,sep="_"),"sce21_quarter_csq_grid.RData",sep="/"))
  sce21 <- subset(sce21,sce21@data$csquares %in% Regiontab1$csquares)
  if(nrow(sce21)!=0){
  sce21@data$area_sqkm <- area(sce21)/10^6
  sce21 <- cbind(sce21, Regiontab1[match(sce21$csquares,Regiontab1$csquares), c("mobeff")])
  colnames(sce21@data)[ncol(sce21@data)] <- "mobeff"
  tab2[13,4] <- round(sum(sce21@data$area_sqkm * sce21@data$mobeff,na.rm=T) / sum(Regiontab1$area_sqkm * Regiontab1$mobeff,na.rm=T) * 100,digits =1)
  } else {
    tab2[13,4] <- "NA"
  }
  
  load(paste(pathdir,paste("2-Data processing/VME_polygons",datacallyear,sep="_"),"sce22_quarter_csq_grid.RData",sep="/"))
  sce22 <- subset(sce22,sce22@data$csquares %in% Regiontab1$csquares)
  if(nrow(sce22)!=0){
  sce22@data$area_sqkm <- area(sce22)/10^6
  sce22 <- cbind(sce22, Regiontab1[match(sce22$csquares,Regiontab1$csquares), c("mobeff")])
  colnames(sce22@data)[ncol(sce22@data)] <- "mobeff"
  tab2[13,5] <- round(sum(sce22@data$area_sqkm * sce22@data$mobeff,na.rm=T) / sum(Regiontab1$area_sqkm * Regiontab1$mobeff,na.rm=T) * 100,digits =1)
  } else {
    tab2[13,5] <- "NA"
  }
  
  
  load(paste(pathdir,paste("2-Data processing/VME_polygons",datacallyear,sep="_"),"sce23_quarter_csq_grid.RData",sep="/"))
  sce23 <- subset(sce23,sce23@data$csquares %in% Regiontab1$csquares)
  if(nrow(sce23)!=0){
  sce23@data$area_sqkm <- try(area(sce23)/10^6)
  sce23 <- cbind(sce23, Regiontab1[match(sce23$csquares,Regiontab1$csquares), c("mobeff")])
  colnames(sce23@data)[ncol(sce23@data)] <- "mobeff"
  tab2[13,6] <- round(sum(sce23@data$area_sqkm * sce23@data$mobeff,na.rm=T) / sum(Regiontab1$area_sqkm * Regiontab1$mobeff,na.rm=T) * 100,digits =1)
  } else {
    tab2[13,6] <- "NA"
  }
  
  
# overlap between mobile fishing area and vme polygons
  SF_scen11 <- st_intersection(s_scen11_dpt,Mobile_fish); SF_scen11 <- st_make_valid(SF_scen11)
  tab2[14,2] <- round(sum(st_area(SF_scen11))/sum(st_area(Mobile_fish)) * 100,digits = 1)
  
  SF_scen12 <- st_intersection(s_scen12_dpt,Mobile_fish); SF_scen12 <- st_make_valid(SF_scen12)
  tab2[14,3] <- round(sum(st_area(SF_scen12))/sum(st_area(Mobile_fish)) * 100,digits = 1)
  
  SF_scen21 <- st_intersection(s_scen21_dpt,Mobile_fish); SF_scen21 <- st_make_valid(SF_scen21)
  tab2[14,4] <- round(sum(st_area(SF_scen21))/sum(st_area(Mobile_fish)) * 100,digits = 1)
  
  SF_scen22 <- st_intersection(s_scen22_dpt,Mobile_fish); SF_scen22 <- st_make_valid(SF_scen22)
  tab2[14,5] <- round(sum(st_area(SF_scen22))/sum(st_area(Mobile_fish)) * 100,digits = 1)
  
  SF_scen23 <- st_intersection(s_scen23_dpt,Mobile_fish); SF_scen23 <- st_make_valid(SF_scen23)
  tab2[14,6] <- round(sum(st_area(SF_scen23))/sum(st_area(Mobile_fish)) * 100,digits = 1)

tab2[,1] <- c("VME polygon description","","VME polygon outcomes","% of fishable domain identified as VME polygon",
              "% of VME polygon protected by existing VME fishery closures",
              "Number of VME polygons and their average areal extent (size)",
              "Number (and average size) of large VME polygons in upper 25<sup>th</sup> percentile of the size distribution",
              "Total area identified under scenario", "Risk to VME*","Fishery consequences",
              "% of effort per year by static gear (400-800m depth) overlapping with VME polygons (average annual effort between 2019 to 2021)",
              "% of fished area (400-800m depth) by static gear overlapping with VME polygons between 2019 to 2021",
              "% Effort represented by SAR of mobile gear (400-800m depth) overlapping with VME polygons (average annual SAR between 2019 to 2021)",
              "% of fished area (400-800m depth) by mobile gear overlapping with VME polygons between 2019 to 2021"
              )

tab2[1,2:6] <- c("All VME habitats, High and Medium VME Index. Low VME Index: only if adjacent to medium or high Index VMEs.",
               "Option 1 + selected VME elements (banks, seamounts, coral mounds, mud volcanoes) associated with any VME records.",
               "All VME habitats, High and Medium VME Index. Low VME Index: only if adjacent to higher index VMEs and Low VME Index in C-squares with low fishing pressure",
               "All VME habitats, High, Medium and Low VME Index excluding C-squares with high fishing pressure (SAR > 0.43)",
               "Scenario 2, Option 1 + selected VME elements (banks, seamounts, coral mounds, mud volcanoes) associated with any VME records.")
               
colnames(tab2) <- c("","Scenario 1 Option 1","Scenario 1 Option 2","Scenario 2 Option 1","Scenario 2 Option 2",
                    "Scenario 1 Option 2 + Scenario 2 Option 1")
